+++ tests/test_detailed_files_rot_timestamp.rs
@@ -11,7 +11,7 @@ fn test_detailed_files_rot_timestamp() {
         .log_to_file()
         .rotate_over_size(2000)
         .o_timestamp(true)
-        .start_reconfigurable()
+        .start()
         .unwrap_or_else(|e| panic!("Logger initialization failed with {}", e));
 
     error!("This is an error message");