+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.6] - 2019-02-26
+
+Deprecate `Logger::start_reconfigurable()`, let `Logger::start()` return a reconfiguration handle.
+
+Add an option to write all logs to nowhere (i.e., do not write any logs).
+
 ## [0.10.5] - 2019-01-15
 
 Eliminate performance penalty for using reconfigurability.